+++ graylog2-server/src/main/java/org/graylog2/rest/resources/RestResourcesModule.java
@@ -18,9 +18,9 @@ package org.graylog2.rest.resources;
 
 import org.graylog.plugins.views.search.engine.monitoring.data.histogram.rest.HistogramResponseWriter;
 import org.graylog.plugins.views.storage.migration.RemoteReindexResource;
+import org.graylog.security.rest.CAResource;
+import org.graylog.security.rest.CertificateRenewalResource;
 import org.graylog2.Configuration;
-import org.graylog2.bootstrap.preflight.web.resources.CAResource;
-import org.graylog2.bootstrap.preflight.web.resources.CertificateRenewalResource;
 import org.graylog2.bootstrap.preflight.web.resources.DataNodeProvisioningResource;
 import org.graylog2.contentstream.rest.ContentStreamResource;
 import org.graylog2.plugin.inject.Graylog2Module;
